requirements.txt
python-gnupg
pycrypto
